Turned on SINGLE precision mode. Potential arrays remain as double in atomicAdd. All kernels give correct potential energies up to the 7th digit.

+++ src/interaction_compute/cuda_downpass.cu
@@ -61,7 +61,7 @@ extern "C"
 void CUDA_Wrapup(int target_xyz_dim, double *potential)
 {
     cudaErr = cudaMemcpy(potential, d_potential,
-                         target_xyz_dim * sizeof(FLOAT), cudaMemcpyDeviceToHost);
+                         target_xyz_dim * sizeof(double), cudaMemcpyDeviceToHost);
     if ( cudaErr != cudaSuccess )
         printf("Device to Host MemCpy failed with error \"%s\".\n", cudaGetErrorString(cudaErr));
